Extract launched effect module
